requirements.txt
altgraph==0.17.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.13.1
fsspec==2023.10.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
openai-whisper==20230918
packaging==23.2
pefile==2023.2.7
psutil==5.9.6
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
PyGetWindow==0.0.9
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
PyRect==0.2.0
pywin32-ctypes==0.2.2
regex==2023.10.3
requests==2.31.0
sounddevice==0.4.6
sympy==1.12
tiktoken==0.3.3
torch==2.1.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
watchdog==3.0.0
Wave==0.0.2
wavio==0.0.8
